Update Phoenix to latest and add --no-ecto option

+++ phoenix/dockstrap.sh
@@ -10,7 +10,7 @@ RUN mix local.hex --force
 RUN mix local.rebar --force
 
 # Install the Phoenix framework
-RUN mix archive.install hex phx_new 1.4.3 --force
+RUN mix archive.install hex phx_new 1.4.4 --force
 
 # Prep node to install via apt
 RUN curl -sL https://deb.nodesource.com/setup_12.x  | bash -
@@ -25,9 +25,6 @@ HEREDOC
 DOCKER_COMPOSE=$(cat <<HEREDOC
 version: "3"
 
-volumes:
-  pg-data:
-
 services:
   app:
     build:
@@ -36,11 +33,14 @@ services:
       - "4000:4000"
     volumes:
       - .:/app
-    depends_on:
-      - postgres
     command:
       mix phx.server
+HEREDOC
+)
 
+POSTGRES_SERVICE=$(cat <<HEREDOC
+    depends_on:
+      - postgres
   postgres:
     image: postgres
     environment:
@@ -49,6 +49,9 @@ services:
       PGDATA: /var/lib/postgresql/data
     volumes:
       - pg-data:/var/lib/postgresql/data
+
+volumes:
+  pg-data:
 HEREDOC
 )
 
@@ -62,7 +65,15 @@ echo lower-case letters with underscores
 
 read -p 'App Name (ex. hello_world): ' APPNAME
 
-echo 'Y' | docker-compose run --no-deps --rm app mix phx.new . --app $APPNAME
+read -p 'use --no-ecto [yN] ' NO_ECTO
+
+if [[ $NO_ECTO =~ [yY](es)* ]]; then
+  echo 'using --no-ecto option'
+  echo 'Y' | docker-compose run --no-deps --rm app mix phx.new . --no-ecto --app $APPNAME
+else
+  echo "$POSTGRES_SERVICE" >> docker-compose.yml
+  echo 'Y' | docker-compose run --no-deps --rm app mix phx.new . --app $APPNAME
+fi
 
 RESULT=$?
 if [ $RESULT -eq 0 ]; then
@@ -72,7 +83,11 @@ if [ $RESULT -eq 0 ]; then
   docker-compose run --rm --no-deps app sed -i 's/localhost/postgres/g' config/dev.exs
   docker-compose run --rm --no-deps -w /app/assets app npm install && node node_modules/webpack/bin/webpack.js --mode development
 
+if [[ $NO_ECTO =~ [yY](es)* ]]; then
+  echo ""
+else
   docker-compose run --rm app mix ecto.create
+fi
 
   # ask to bring it up now
   RESULT=$?
